requirements.txt
Pillow
bottle
requests>=1.0
pyproj
six
